Let's polish the use of Streams in TaskList.
Let's also refactor the different TaskList related classes and methods to a separate package.
